Implement Manager class
I have captured the Manager tasks in a Manager class.  Manager objects
are used by iteratively calling Manager::iterate() in an event loop.  I
have implemented the Manager conceptually as a finite state machine,
where the Manager can be in one of three states:

* Idle: No Worker process running, waiting for input from the Master
* Busy: Worker process running, waiting for input from Worker
* Terminated: No Worker process running and not waiting for any input

An idle Manager transitions to a busy Manager when it receives an input
message from the Master.  The associated action is to launch the
appropriate Worker process.  Similarly, a busy Manager transitions to an
idle Manager when its Worker finishes.  The associated action is to send
the result of the Worker process to the Master.

In addition, the Manager can also receive two types of signals from the
Master:

* TERMINATE_MANAGER_SIGNAL: Terminate Manager
* TERMINATE_PROCESS_SIGNAL: Terminate Worker process

When TERMINATE_MANAGER_SIGNAL is received, the Manager will transition
to the terminated state.  If the Manager is busy, it will first
terminate its Worker process.

When TERMINATE_PROCESS_SIGNAL is received, the busy Manager will
terminate its Work process, send a CANCEL result to the Master, and
become idle.  Sending TERMINATE_PROCESS_SIGNAL to an idle Manager is an
error.

It is illegal to call Manager::iterate() on a terminated Manager.  The
state can be checked with Manager::getState().  Alternatively, if
Manager::isActive() is true, the Manager is either in the idle or busy
state.

Finally, when the program receives a SIGINT or SIGTERM signal, the
boolean flag m_p_program_terminated is switched on.  This will cause the
Manager to terminate as if it received TERMINATE_MANAGER_SIGNAL.