Send output message to Master from Managers
In order to increase the flexibility of the Pakman parallel framework,
the Managers now relay the entire output message from Workers to the
Master.  This is opposed to Managers preprocessing the output from
Workers and sending back predefined results.

In order to handle cancelled processes, I gave the Manager the
capability to send signals to the Master.  Thus, I renamed the current
tag SIGNAL_TAG to MASTER_SIGNAL_TAG and created the new tag
MANAGER_SIGNAL_TAG.