Implement Abstract and ABCRejection Master classes
I have captured the Master tasks in a Master class hierarchy. Master
objects are used by iteratively calling Master::iterate() in an event
loop.  The main objective of the Master class conceptually is to manage
a "Master queue" that is composed of three regions:

* Finished region: This is the front part of the Master queue, where
finished tasks accumulate.

* Busy region: This is the middle part of the Master queue.  It contains
tasks that the Master has sent to Managers to be completed.  Since tasks
are not necessarily completed in the order that the Master sent them
out, this region may contain finished tasks.  The boundary between the
finished and busy region is demarcated by the first unfinished task of
the Master.

* Pending region: This is the last part of the Master queue and it is
composed of tasks that have not yet been sent out to Managers to be
completed.  The boundary between the busy and pending region is
demarcated by the first pending task.

When Master::iterate() is called, it listens to Manager messages, which
converts busy tasks to finished tasks, moving the boundary between
finished and busy tasks to the back of the Master queue.  It also sends
out pending tasks to idle Managers, converting pending tasks to busy
tasks and pushing the boundary between finished and pending tasks
backwards.

Moreover, within Master::iterate(), the Master object calls a pure
virtual function Master::processTasks().  The purpose of this function
is to let derived classes manipulate the Master queue in one of two
ways:

* Using the protected function Master::queueTask(), derived classes can
queue new tasks, adding them to the pending region of the Master queue.

* Using the protected function Master::popFinishedTasks(), derived
classes can pop finished tasks from the finished region of the Master
queue.

Hence, when implementing ABC or other algorithms, the implementer is
conceptually interacting with a queue of tasks, to which it can push
tasks and from which it can pop finished tasks.  The internal
implementation of Master then automatically distributes the tasks to the
pool of Managers so that they can be processed in parallel.  This
greatly simplifies the task of implementing a new ABC algorithm.

In addition, I have also added the possibility of completely flushing
the Master queue.  This is necessary for some algorithms, such as the
ABC sequential Monte Carlo algorithm.  In order to do this, the derived
class calls Master::terminateWorkers() within Master::processTasks().
On the next iteration, the Master will then transition to a flushing
state, in which it stops calling Master::processTasks() and listens to
Manager results and cancellation signals until all Managers are idle.
Only when all Managers are confirmed to be idle, will the Master return
to its normal state and continue to call Master::processTasks().

When the algorithm has finished, the derived class calls
Master::terminateMasterAndManagers() in Master::processTasks() in order
to indicate that it has finished.  The Master will then transition to a
terminated state.  The public Master::isActive() function will then
return false, and it is then illegal to call Master::iterate().

Finally, I have implemented the ABC rejection algorithm with the derived
class ABCRejectionMaster.