Implement Master and Controller class hierarchies
In order to improve the modularity of the code, I separated out the
tasks of the "Master" and the "Controller" into separate class
hierarchies.  The Master classes are responsible for processing Tasks
pushed by the Controller classes.  Thus, the Controller classes contain
purely the logic of the ABC or other algorithms, completely
independently of how the tasks are executed or parallelised.  Meanwhile,
the Master classes provide the mechanics for executing the tasks.  This
also means that no separate code needs to be written for a serial or
parallelised ABC algorithm, since the SerialMaster (to be implemented)
and MPIMaster can both accept the same Controller.

I have put the old master routines into the original
methods/mpi/master.{h,cc} files so they do not interfere with the
Controller class hierarchy.  As I implement Controller classes for the
smc, sweep and mcmc algorithms, these files will eventually disappear.

Lastly, I have created a specific methods/common.h file to contain the
constants that are not MPI-specific.