Change Linux-specific pipe2() to portable pipe()
I used pipe2 before to pass the O_CLOEXEC flag (close-on-exec) to avoid
leaking file descriptors in a multithreaded environment.  Since we have
eliminated multithreading from our program, we no longer need to do
this.  In addition, using pipe() is more portable as it is covered by
the POSIX standard, while pipe2() is a Linux-specific system call.