Record error code from Worker
I have implemented the communication of an error code from the Worker to
the Manager, who then relays it to the Master.  The Controller can query
the error code of a finished Task by calling
AbstractMaster::TaskHandler::getErrorCode().

In the case of a forked Worker, the error code is the exit status of the
forked process.

In the case of an MPI Worker, the error code is communicated with the
tag WORKER_ERROR_CODE_TAG by the Worker to the Manager.