Implement C header for running serial MPI Worker
In accordance with the communication between the Manager and MPI Worker,
I have created the header include/pakman_mpi_worker.h that provides the
function

int pakman_run_mpi_worker(int *(simulator)(const char *input_string,
    char **p_output_string), int flags)

This function takes a function pointer to a function that takes in an
input string and returns an output string, as well as an error code.
This is the function is provided by the user to perform the simulation.

The flags variable control whether MPI is initialized or finalized
inside the function pakman_run_mpi_worker.  It also controls whether the
Worker is persistent or not.  This is a feature that will be implemented
in the future.

Currently this will only work for MPI Workers that are serial, e.g. the
MPI Worker consists of a single MPI process.