Implement persistent MPI Worker
I have created a persistent MPI Worker, which is a spawned MPI process
that does not exit after performing a task.  Instead, it waits for the
next task.  It continues doing to so until it receives the signal
TERMINATE_WORKER_SIGNAL from its Manager.

In order to accomplish this, I have slightly altered the semantics of
the communication between the Master and Manager.  Instead of sending a
TERMINATE_WORKER_SIGNAL, the Master now sends a FLUSH_WORKER_SIGNAL to
the Manager when it is flushing its queues.  "Flushing" a Worker is
achieved by calling the destructor of the corresponding
AbstractWorkerHandler.  Terminating is achieved by calling
AbstractWorkerHandler::terminate().

For non-persistent Workers, flushing implies process termination.  For
persistent Workers, flushing does not imply process termination.