Implement static termination for persistent workers
I have implemented the static function
PersistentMPIWorkerHandler::terminatePersistent() in order to terminate
any persistent Workers that remain after their respective Managers have
terminated.

This situation occurs when the Manager is idle at the time of receiving
a termination signal from the Master.  In that case, the Manager has no
instance of PersistentMPIWorkerHandler to call terminate() on, so the
persistent Worker outlives its Manager.