Implement LongOptions and Arguments classes
These classes facilitate parsing command-line arguments when the options
to parse are determined at runtime.  This is relevant to Pakman, as the
accepted arguments are determined by the choice of master and
controller.

The LongOptions class is agnostic with regards to the source of options.
The Arguments class then uses the LongOptions class to process and store
the command-line arguments.