Add scaling test for single node
I created the directory /scaling to run a scaling test that benchmarks
pakman with a compute-intensive simulator on an increasing number of
parallel processes (powers of 2), until the total number of cores in the
system has been reached.

Note that I have used the CMake module "ProcessorCount" for this, which
in some cases (at least in mine) displays the number of logical cores as
opposed to the number of physical cores.