Implement simple ABC rejection inference test
I have implemented a simple test of the ABC rejection inference
algorithm.

In order to obtain a deterministic and non-trivial list of parameters, I
wrote the script increment-and-print-number.sh, which takes as its
argument a "number file" containing an integer number in plaintext.
Every time the script is run, it increments the number read from the
number file, prints it out and stores the incremented number back into
the number file.  Repeated runs of this script thus result in a sequence
of increasing numbers.

The script accept-if-epsilon-plus-parameter-is-even.sh pretty much does
what it says on the tin.  It reads epsilon and a parameter from its
standard input, and accepts if their sum is even and rejects otherwise.

By using increment-and-print-number.sh and
accept-if-epsilon-plus-parameter-is-even.sh as the prior_sampler and
simulator, respectively, to pakman, the accepted parameter list should
be a list of even or odd numbers, when the epsilon is even or odd,
respectively.