Implement simple ABC SMC inference test
I have implemented a simple test of the ABC SMC inference algorithm.

The simulator and prior sampler are the same as in the simple ABC
rejection inference test.

The script perturber.sh accepts the epsilon list as a command-line
argument to check whether the parameters it receives as input are valid
for the given generation.  If so, it "perturbs" the parameter by
incrementing it.

The script prior-pdf.sh only checks whether it receives the correct
number of lines as input and always prints 1.

The script perturbation-pdf.sh also accepts the epsilon list as a
command-line argument to perform the same check as perturber.sh.
Moreover, it also checks whether the perturbed parameter is in fact a
parameter of the previous generation that is incremented.