Add simulator protocol
I have created a subdirectory "interface" that will deal with the
interface between pakman and external components and user input.  One of
its main components is the protocols.cc file, which defines how pakman
communicates with external components, aka protocols.

The first protocol to be added is the simulator protocol.