Move input struct into Controller class
Previously, the input object that was generated from the input file and
passed to Controller classes was defined as a input_t struct in
core/types.h.  A conflict between input_t types was avoided by using
namespaces.  This approach originated from when pakman was still
compiled as a separate binary for each combination of master and
controller.

Now, I have a defined a nested Input struct in each Controller class to
perform the same function.  I have also removed core/types.h as it no
longer served any purpose.