Remove non-persistent MPI simulator functionality
When testing Pakman with CTest, there is a chance that users will run
into errors for the non-persistent MPI simulator tests.  This is because
in this case Pakman needs to repeatedly spawn MPI processes, for which
MPI support is the weakest.

Such failures highlight the danger of including tests with of
non-persistent MPI simulators, since it may give users the impression
that their build of Pakman is dysfunctional when it can be perfectly
used for other types of simulators.

The advantage of keeping non-persistent MPI simulators around is that it
gives the user the ability to explicitly spawn a new instance of the MPI
simulator every time a simulation is issued.  This ensures that no
program state survives in subsequent simulations.

With persistent MPI simulators, the burden is on the user to write a
function that does not keep state between function calls, i.e.  does not
use static or global variables and properly closes all external files.

In the spirit of "less is more", I have decided to cut out
non-persistent MPI simulator functionality from Pakman completely.  I
believe having two types of MPI simulators will only serve to confuse
users.  Moreover, there is no difference in how to write MPI simulator
functions to give to Pakman MPI workers.  It also gives more power to
the user, performs better than non-persistent MPI simulators and
eliminates the risk of exposing the user to the weak MPI support of
dynamic process management.