Reimplement cell growth example with ABC rejection
I have reimplemented the epithelial cell growth example so that the
example code is contained completely within in the repository, rather
than being in a separate directory.

The user needs to have a version of Chaste on their computer, and have
the environment variable CHASTE_DIR be equal to the Chaste root
directory.  The optional environment variable CHASTE_BUILD_DIR tells
CMake in which directory to build Chaste.  If this variable is not set,
CMake will create a Chaste build tree in the Pakman build tree.

Alternatively, CHASTE_DIR and CHASTE_BUILD_DIR can be specified as CMake
variables when running the cmake command.  If specified as CMake
variables, they will have precedence over the environment variables.