Narrow domain of prior distribution
After fixing the meaning of tcycle in the epithelial cell growth model,
it is more appriopriate to narrow down the range of the prior
distribution.  It is not possible to have a tcycle of zero, so the lower
bound was set on tcycle = 5.  The upper bound was changed to 25 and the
observed data was changed to 13.  Finally, the epsilon series is now
4,3,2,1,0 and the perturber has a standard deviation of 1.