Add utils/run-mpi-worker executable
I have implemented the utility executable `run-mpi-worker`, which is
used to run an MPI Worker exactly once.  It is essentially a way to
convert an MPI simulator to a Standard simulator.  The executable
run-mpi-worker takes two lines from standard input and passes it on to
the MPI simulator command specified on its command-line arguments.  It
then reads the output and prints it to standard output.