Implement Command::isExecutable check
Previously, if a call to exec failed and the flag --ignore-errors was
set, Pakman would not give an error when the executable file for a child
process was missing.  This is because throwing an error after exec would
not affect the execution of Pakman because the nonzero exit code would
be ignored.

Now, I have added a check at the beginning of system_call functions to
see if the executable file in the given Command can be found and is
executable using the system call access.